requirements.txt
# NSE Intraday Cross-Sectional Mean Reversion
# Python dependencies

# Data
yfinance>=0.2.36
pyarrow>=15.0.0

# Core
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# Config
pyyaml>=6.0

# Timezone
pytz>=2024.1

# Testing
pytest>=8.0.0
